refactor: form 配下のサブモジュールから不要な models 階層を除去
各サブモジュール (answer/comment/message/message_thread/question) が
xxx.rs に `pub mod models;` だけを持ち、実体が xxx/models.rs にあるという
1階層深いラッパー構造になっていた。同じディレクトリに service / repository
など複数ファイルを並べる予定がない以上、この階層は名前空間を冗長にする
だけなので、xxx/models.rs を xxx.rs にフラット化した。

あわせて、ActiveForm / FormLabel の AuthorizationGuardDefinitions impl に
付いていた `# Examples` doctest 8 個を削除した。中身は administrator と
standard_user を毎回組み立てて `is_administrator(actor)` 相当の振る舞いを
再宣言するだけのボイラープレートで、各 impl 本体に対して情報量が無く、
form/models.rs の見通しを大きく損ねていた。policy は日本語の説明文として
残し、必要なら既存の tests モジュールでカバーする方針。
