+++ server/domain/src/form/message.rs
@@ -1,1 +1,128 @@
-pub mod models;
+use chrono::{DateTime, Utc};
+use derive_getters::Getters;
+use domain_derive::UnsafeFromRawParts;
+use errors::domain::DomainError;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    form::message_thread::MessageThread,
+    types::authorization_guard::{
+        AuthorizationRole, BelongsTo, Delete, GuardedBy, ParentGuarded, Update,
+    },
+    user::models::{Actor, Role::Administrator, User, UserId},
+};
+
+pub type MessageId = types::Id<Message>;
+
+impl AuthorizationRole for Message {
+    type Role = ParentGuarded<MessageThread>;
+}
+
+#[derive(UnsafeFromRawParts, Getters, Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub struct Message {
+    id: MessageId,
+    sender_id: UserId,
+    body: String,
+    timestamp: DateTime<Utc>,
+}
+
+impl Message {
+    /// [`Message`] の生成を試みます。
+    ///
+    /// 以下の場合に失敗します。
+    /// - [`body`] が空文字列の場合
+    ///
+    /// # Examples
+    /// ```
+    /// use domain::{
+    ///     form::message::Message,
+    ///     user::models::UserId,
+    /// };
+    /// use uuid::Uuid;
+    ///
+    /// let user_id: UserId = Uuid::new_v4().into();
+    ///
+    /// let success_message = Message::try_new(user_id, "test message".to_string());
+    /// let message_with_empty_body = Message::try_new(user_id, "".to_string());
+    ///
+    /// assert!(success_message.is_ok());
+    /// assert!(message_with_empty_body.is_err());
+    /// ```
+    pub fn try_new(sender_id: UserId, body: String) -> Result<Self, DomainError> {
+        if body.is_empty() {
+            return Err(DomainError::EmptyMessageBody);
+        }
+
+        Ok(Self {
+            id: MessageId::new(),
+            sender_id,
+            body,
+            timestamp: Utc::now(),
+        })
+    }
+
+    pub fn update_body(self, body: String) -> Result<Self, DomainError> {
+        if body.is_empty() {
+            return Err(DomainError::EmptyMessageBody);
+        }
+
+        Ok(Self { body, ..self })
+    }
+}
+
+impl BelongsTo<MessageThread> for Message {
+    fn belongs_to(&self, parent: &MessageThread) -> bool {
+        parent
+            .messages()
+            .iter()
+            .any(|message| message.id() == self.id())
+    }
+}
+
+impl GuardedBy<MessageThread, Update> for Message {
+    fn is_allowed_for(&self, _parent: &MessageThread, actor: &Actor) -> bool {
+        matches!(actor, Actor::User(User::ActiveUser(user)) if self.sender_id() == user.id())
+    }
+}
+
+impl GuardedBy<MessageThread, Delete> for Message {
+    fn is_allowed_for(&self, _parent: &MessageThread, actor: &Actor) -> bool {
+        matches!(
+            actor,
+            Actor::User(User::ActiveUser(user))
+                if self.sender_id() == user.id() || user.role() == &Administrator
+        )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        form::answer::AnswerId,
+        types::authorization_guard::{AuthorizationGuard, Update},
+        user::models::{ActiveUser, Role},
+    };
+    use uuid::Uuid;
+
+    #[test]
+    fn message_update_requires_message_to_belong_to_thread() {
+        let user_id: UserId = Uuid::new_v4().into();
+        let actor = Actor::from(ActiveUser::new(
+            "sender".to_string(),
+            user_id,
+            Role::StandardUser,
+        ));
+        let answer_id: AnswerId = Uuid::new_v4().into();
+        let thread = MessageThread::new(answer_id, user_id)
+            .add_message(Message::try_new(user_id, "owned message".to_string()).unwrap());
+        let foreign_message = Message::try_new(user_id, "foreign message".to_string()).unwrap();
+        let allowed_thread = AuthorizationGuard::<_, Update>::from(thread)
+            .try_update(actor)
+            .unwrap();
+
+        let result = allowed_thread.authorize_update(foreign_message);
+
+        assert!(result.is_err());
+    }
+}
