+++ server/infra/resource/src/database/forms/comment.rs
@@ -4,8 +4,8 @@ use crate::{
 };
 use async_trait::async_trait;
 use domain::form::{
-    answer::models::AnswerId,
-    comment::models::{Comment, CommentId},
+    answer::AnswerId,
+    comment::{Comment, CommentId},
 };
 use errors::infra::InfraError;
