+++ server/presentation/src/handlers/form/answer_label_handler.rs
@@ -7,7 +7,7 @@ use axum::{
     response::{IntoResponse, Response},
 };
 use domain::{
-    form::answer::models::{AnswerId, AnswerLabelId},
+    form::answer::{AnswerId, AnswerLabelId},
     repository::Repositories,
     user::models::ActiveUser,
 };
