+++ server/presentation/src/handlers/form/message_handler.rs
@@ -5,7 +5,7 @@ use axum::{
     response::IntoResponse,
 };
 use domain::{
-    form::{answer::models::AnswerId, message::models::MessageId},
+    form::{answer::AnswerId, message::MessageId},
     repository::Repositories,
     user::models::ActiveUser,
 };
